# docs/swagger.yaml
basePath: /v2
definitions:
  domain.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  domain.Pet:
    properties:
      category:
        $ref: '#/definitions/domain.Category'
      id:
        type: integer
      name:
        type: string
      photoUrls:
        items:
          type: string
        type: array
      status:
        type: string
      tags:
        items:
          $ref: '#/definitions/domain.Tag'
        type: array
    type: object
  domain.Tag:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Petstore API
  version: "1.0"
paths:
  /pet:
    post:
      consumes:
      - application/json
      description: Add a new pet to the store
      parameters:
      - description: Pet object that needs to be added
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/domain.Pet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Pet'
      security:
      - OAuth2Application:
        - write
      summary: Create a pet
      tags:
      - pet
  /pet/{petId}:
    get:
      consumes:
      - application/json
      description: Returns a single pet
      parameters:
      - description: ID of pet to return
        in: path
        name: petId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Pet'
      security:
      - ApiKeyAuth: []
      summary: Get pet by ID
      tags:
      - pet
swagger: "2.0"